CLOUDP-331821: Refactor validations.js split into multiple files
