+++ tools/spectral/ipa/rulesets/functions/IPA104GetResponseCodeShouldBe200OK.js
@@ -6,7 +6,7 @@ import {
   isSingleResourceIdentifier,
   isSingletonResource,
 } from './utils/resourceEvaluation.js';
-import { checkResponseCodeAndReturnErrors } from './utils/validations.js';
+import { checkResponseCodeAndReturnErrors } from './utils/validations/checkResponseCodeAndReturnErrors.js';
 
 const RULE_NAME = 'xgen-IPA-104-get-method-response-code-is-200';
 const ERROR_MESSAGE =
