+++ tools/spectral/ipa/rulesets/functions/IPA106CreateMethodRequestHasNoReadonlyFields.js
@@ -7,7 +7,7 @@ import {
 import { resolveObject } from './utils/componentUtils.js';
 import { hasException } from './utils/exceptions.js';
 import { collectAdoption, collectAndReturnViolation, collectException } from './utils/collectionUtils.js';
-import { checkForbiddenPropertyAttributesAndReturnErrors } from './utils/validations.js';
+import { checkForbiddenPropertyAttributesAndReturnErrors } from './utils/validations/checkForbiddenPropertyAttributesAndReturnErrors.js';
 
 const RULE_NAME = 'xgen-IPA-106-create-method-request-has-no-readonly-fields';
 const ERROR_MESSAGE = 'The Create method request object must not include input fields (readOnly properties).';
